Move HTTP Exception handlers to another class for better readability
